Set defaults for newly opened handles and cleanup logging.
Make sure that when new handles are opened we are in autocommit mode, the time
zone is set to UTC, and set the mysql mode to be more strict with regards to
mysql bugs in date handling etc.

Cleanup the error logging using new Dbi_LogException() call.

Log the host and server version when opening a new handle.